fix: Correcting index signatures
